+++ server/notebooks/serializers.py
@@ -61,5 +61,6 @@ class NotebookRevisionDetailSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = NotebookRevision
-        fields = ("id", "title", "created", "content")
+        fields = ("id", "title", "created", "content", "is_draft")
         write_only_fields = "notebook_id"
+        read_only_fields = ["is_draft"]