+++ server/tests/test_notebook_api.py
@@ -39,13 +39,17 @@ def test_notebook_detail(client, test_notebook):
             "created": get_rest_framework_time_string(initial_revision.created),
             "id": initial_revision.id,
             "title": initial_revision.title,
+            "is_draft": False,
         },
     }
 
     # add another revision, make sure all return values are updated
     # appropriately
     new_revision = NotebookRevision.objects.create(
-        notebook=test_notebook, title="Second revision", content="*updated fake notebook content*"
+        notebook=test_notebook,
+        title="Second revision",
+        content="*updated fake notebook content*",
+        is_draft=False,
     )
     resp = client.get(reverse("notebooks-detail", kwargs={"pk": test_notebook.id}))
     assert resp.status_code == 200
@@ -59,6 +63,7 @@ def test_notebook_detail(client, test_notebook):
             "created": get_rest_framework_time_string(new_revision.created),
             "id": new_revision.id,
             "title": new_revision.title,
+            "is_draft": False,
         },
     }
 
@@ -97,6 +102,7 @@ def test_create_notebook_logged_in(fake_user, client, notebook_post_blob, specif
             "created": get_rest_framework_time_string(notebook_revision.created),
             "id": notebook_revision.id,
             "title": notebook_revision.title,
+            "is_draft": False,
         },
     }
     # should have a first revision to go along with the new notebook